+++ embedded-hal-bus/src/spi/shared.rs
@@ -1,4 +1,4 @@
-use embedded_hal::delay::DelayUs;
+use embedded_hal::delay::DelayNs;
 use embedded_hal::digital::OutputPin;
 use embedded_hal::spi::{ErrorType, Operation, SpiBus};
 
@@ -15,7 +15,7 @@ pub fn transaction<Word, BUS, CS, D>(
 where
     BUS: SpiBus<Word> + ErrorType,
     CS: OutputPin,
-    D: DelayUs,
+    D: DelayNs,
     Word: Copy,
 {
     cs.set_low().map_err(DeviceError::Cs)?;